refactor: reorganize imports across services for consistency
